bot_requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
dateparser==1.1.0
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
numpy==1.21.4
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
six==1.16.0
TA-Lib==0.4.22
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ujson==4.3.0
urllib3==1.26.7
websocket-client==1.2.1
websockets==9.1
yarl==1.7.2